LOAN-49: apply neccessary changes to integrate with lecturer api - BK

+++ Src/Core/LoansComparer.CrossCutting/DTO/LoaningBank/GetOfferResponse.cs
@@ -8,7 +8,7 @@ namespace LoansComparer.CrossCutting.DTO.LoaningBank
         public string Id { get; set; } = default!;
 
         [JsonPropertyName("requestedValue")]
-        public int LoanValue { get; set; }
+        public float LoanValue { get; set; }
 
         [JsonPropertyName("requestedPeriodInMonth")]
         public short LoanPeriod { get; set; }
